refactor(server): extract SQL/result interfaces into server/sqlcore so flightclient is duckdb-free
Step 6 of the binary-split plan. Carves the SQL/result interfaces and
two pure helpers out of server/ into a new server/sqlcore subpackage so
the Flight client can talk to those abstractions without importing the
rest of server/ (which still links libduckdb).

Symbols moved:

  - Interfaces: RowSet, ExecResult, RawConn, ColumnTyper, QueryExecutor
    (from server/executor.go)
  - IsEmptyQuery + the previously private isEmptyQuery /
    stripLeadingComments helpers (from server/conn.go and
    server/exports.go) — now sqlcore.IsEmptyQuery and
    sqlcore.StripLeadingComments
  - OTELGRPCClientHandler (from server/otelgrpc_filter.go)

Backward compatibility preserved via type aliases and re-export `var`s
in server/executor.go and a new server/sqlcore_aliases.go. Existing
references to server.RowSet / server.QueryExecutor / server.IsEmptyQuery
/ server.OTELGRPCClientHandler / etc. compile unchanged.

The Flight client (server/flightclient) is the first consumer to drop
its server import entirely. After this PR:

  go list -deps ./server/sqlcore | grep duckdb-go     # empty
  go list -deps ./server/flightclient | grep duckdb-go # empty

This means a hypothetical cmd/duckgres-controlplane that imports
flightclient (and arrowmap, auth, sysinfo, tlscert, ducklake from
earlier PRs) would not link libduckdb at all. The control plane's
remaining duckdb-go linkage comes from its server import for the
larger Server / clientConn / Config surface, which PR #7+ tackles.

LocalExecutor / PinnedExecutor / LocalRowSet stay in server/executor.go
intentionally — they're standalone-mode-only adapters around *sql.DB
and aren't needed by flightclient or the control plane.

Verified:
  - go build ./... clean
  - go build -tags kubernetes ./... clean
  - go test -short ./server/sqlcore/... ./server/flightclient/...
    ./server/... ./controlplane/... — all green (existing
    isEmptyQuery / stripLeadingComments tests in conn_test.go pass via
    the unexported wrappers)
  - go list -deps ./server/sqlcore | grep duckdb-go is empty
  - go list -deps ./server/flightclient | grep duckdb-go is empty
