+++ server/executor.go
@@ -4,52 +4,21 @@ import (
 	"context"
 	"database/sql"
 	"os"
+
+	"github.com/posthog/duckgres/server/sqlcore"
 )
 
-// ColumnTyper provides type name information for a database column.
-// *sql.ColumnType satisfies this interface.
-type ColumnTyper interface {
-	DatabaseTypeName() string
-}
-
-// RowSet represents a set of rows from a query result.
-type RowSet interface {
-	Columns() ([]string, error)
-	ColumnTypes() ([]ColumnTyper, error)
-	Next() bool
-	Scan(dest ...any) error
-	Close() error
-	Err() error
-}
-
-// ExecResult represents the result of a non-query execution.
-type ExecResult interface {
-	RowsAffected() (int64, error)
-}
-
-// RawConn provides access to the underlying driver connection.
-// *sql.Conn satisfies this interface.
-type RawConn interface {
-	Raw(func(any) error) error
-	Close() error
-}
-
-// QueryExecutor abstracts database query execution, allowing both local (*sql.DB)
-// and remote (Arrow Flight SQL) backends.
-type QueryExecutor interface {
-	QueryContext(ctx context.Context, query string, args ...any) (RowSet, error)
-	ExecContext(ctx context.Context, query string, args ...any) (ExecResult, error)
-	Query(query string, args ...any) (RowSet, error)
-	Exec(query string, args ...any) (ExecResult, error)
-	ConnContext(ctx context.Context) (RawConn, error)
-	PingContext(ctx context.Context) error
-	Close() error
-
-	// LastProfilingOutput returns the JSON profiling output from the last
-	// executed query, or "" if profiling is not enabled or not available
-	// (e.g. Flight SQL mode where the query ran on a remote worker).
-	LastProfilingOutput() string
-}
+// The SQL/result interfaces moved to server/sqlcore so the Flight client
+// and other duckdb-free callers can implement them without importing
+// server. The aliases below preserve the old server.X spellings for the
+// dozens of references inside this package and elsewhere.
+type (
+	ColumnTyper   = sqlcore.ColumnTyper
+	RowSet        = sqlcore.RowSet
+	ExecResult    = sqlcore.ExecResult
+	RawConn       = sqlcore.RawConn
+	QueryExecutor = sqlcore.QueryExecutor
+)
 
 // LocalExecutor wraps *sql.DB to implement QueryExecutor for local DuckDB access.
 type LocalExecutor struct {
