+++ server/sqlcore/otel.go
@@ -1,4 +1,4 @@
-package server
+package sqlcore
 
 import (
 	"strings"
